+++ src/router/rt_list.rs
@@ -1,6 +1,6 @@
 //! Ouverture d'une view
 //!
-use crate::sqlic::sql_crud::crud_read_all;
+use crate::cruder::sql_crud::crud_read_all;
 // use crate::sqlic::sql_utils::querlite;
 use crate::{
     // lexic::lex_table::{self, Element},
@@ -22,6 +22,7 @@ use std::{
     // collections::HashMap,
     sync::atomic::Ordering
 };
+use crate::service;
 
 // cuerl http://0.0.0.0:8080/
 #[get("/list/{appid}/{tableid}/{viewid}")]
@@ -32,6 +33,9 @@ pub async fn list(
     // msg: Option<ReqData<servic::sr_data::Msg>>,
 ) -> Result<impl Responder> {
     // log::info!("Session {:?} {:?} {:?}", session.status(), session.entries(), path);
+    let mut messages = session.get::<Vec<service::Message>>("messages")?.unwrap();
+    messages.push(service::Message::new("list:Tout va bien", service::MESSAGE_LEVEL_INFO));
+
     let (appid, tableid, viewid) = path.into_inner();
     let ptr = data.plexic.load(Ordering::Relaxed);
     let apps = unsafe { &(*ptr).applications.clone() };
@@ -39,7 +43,7 @@ pub async fn list(
 
     let mut rowsel = Vec::new();
 
-    crud_read_all(&session, &data.db, &data.dblite, app, &tableid, &viewid ,&"".to_string(), &mut rowsel).await;
+    crud_read_all(&data.db, &data.dblite, app, &tableid, &viewid ,&"".to_string(), &mut rowsel, &mut messages).await;
 
     // TESTS SQLITE
     // let _reslite: HashMap<String, String> = match querlite(&data.dblite,
@@ -65,6 +69,7 @@ pub async fn list(
     // }
 
     let mut context = tera::Context::new();
+    context.insert("messages", &messages);
     context.insert("portail", unsafe { &(*ptr).portail });
     context.insert("application", &app);
     context.insert("tableid", &tableid);