+++ src/servic/mod.rs
@@ -1,6 +1,0 @@
-
-pub mod sr_data;
-
-// use self::servic_data::AddMsg;
-
-pub mod sr_redirect;